Removing former entry points.
Now that we have a single entry point for InVEST, it doesn't really make sense
to keep these around.  It's often easier (and requires fewer characters) to
just type the cli name and the model name.

Also, since we now have a python script entry point, it's even easier to type this
since your python environment should have the invest binary if natcap.invest
is installed.