re #BITBUCKET-3490 Adding a work around for getting geomorphology values.
- wrote a function that converts a raster to a point shapefile. Function uses the center of a pixel to set geometry coordinates.
- wrote a function that gets the shortest distances between point features and another set of shapefile features, by using an rtree for indexing. A field from the the other shapefile closest feature to a point is added to the point. The point inherits a field from the closest shapefile feature.

Using the above two functions, convert the shore pixels to a point shapefile. Get the closest distance from those points to the geomorphology layer and add the RANK field from the geomorph layer to the point. Burn the Rank values from the points back onto a raster. This should satisfy the problem of getting the geomorphology layer reflected onto the shore pixels.