Virtual environments will now always have nose installed.
This change resolves the issue encountered on Windows computers, where
pygeoprocessing could not be imported.  The issue was due to the
inadvertent use of the nose installed to the system's site-packages
since a nose was not being installed to the virtualenv after the split
of the two requirements.txt files.  In this case, the system
site-packages' nose was attempting to import pygeoprocessing from system
site-packages, which was failing with an importerror.

RE: #BITBUCKET-3506