Allowing the user to install an advanced data zipfile from CLI
The user now may provide the /DATAZIP= command-line flag, pointing to a
LOCAL zipfile.  At the moment, this may not live on a network drive.
If this option is provided, it uses the 'advanced' option on the front
pane of the UI if the UI is being presented.  If the installer is being
run in silent mode (with the /S flag), the zipfile is blindly unzipped
with the rest of the install.

Also, removing a .swp file that shouldn't be there.

RE:#BITBUCKET-3520