Shrinking size of iterblock chunks.
A user wrote to me reporting a MemoryError on the windows builds with
only 3 timesteps and rasters of dimensions 1305x2018.  This is a problem
because the current implementation has a bit of a multidimensional
matrix proliferation problem, where a whole lot of matrices (and in many
cases, matrices with a 3rd dimension where its size is dependent on the
number of transition years) are kept in memory all at the same time.
In this case, iterblocks using up to 2**20 matrix elements is causing
the implementation to exhaust the available addressable memory space.

RE:#BITBUCKET-3640