Restoring imports to the old-style internal package imports.
As far as the sphinx documentation is concerned, this is the easiest
import to deal with.

When using the `from __future__ import absolute_import` style of
imports, the import always fails.  But then when being extra explicit,
such as with `from natcap.invest.pygeoprocessing_0_3_3 import
geoprocessing_core`, the import fails as well because `invest` hasn't
yet been successfully imported.

Using this import type gets around the whole issue.  RE:#BITBUCKET-3649