Implementing validation for crop production model.
This implementation includes a slightly refactored approach to
validation that uses a context manager to reduce the verbosity of the
validation work itself.  The Crop Production -- Percentile model has
been refactored to use the new validation API to allow the current sets
of tests to pass.  RE:#BITBUCKET-3610 RE:#BITBUCKET-3597