Changing how model validation errors are visualized.
When validation errors are present, the model form will now update the
icon of the run button to be a red X, and change it back to the green
enter icon when validation errors are fixed.  If the run button is
pressed when there are validation errors, the validation error dialog
for the whole model is shown.  RE:#BITBUCKET-3610