Removing drag-n-drop of scenarios onto the Model window.
The drag-n-drop works great for the Model window itself, but then it
goes away when we try to drop it onto the Form object that's embedded
into the Model window.  As nice an idea as it was, it ends up being much
easier to just file --> open the file.  RE:#BITBUCKET-3610