Adding an option to clear local settings.
The InVEST settings dialog is for global parameters, so it doesn't
really make sense to put local settings there.  I've added this option
as a menu item under a new 'Edit' menu, but I'm not 100% happy with it
there, either.  It works, so that's good enough for now.  RE:#BITBUCKET-3680