Correcting an issue with phony packages that was forcing the UG to rebuild.
It turns out that having the UG and SVN repositories be PHONY targets
implies that they will always have their recipes reexecuted.  This
then updates the timestamps on their directories, which then triggers
the UG to be rebuilt, which is not something we want.  Rather than
having the repos be PHONY, make should be called with the -B flag
to unconditionally rebuild all targets.  So, `make -B fetch`
would do the trick.  RE:#BITBUCKET-3852