Update issue #27 adding coordinate transform func
Adding coordinate transformation function to utils.py so that all InVEST
models that handle their own transformations can now have a single
entry point for it. This function is compatible with gdal 3 but
NOT gdal 2. This function sets unprojected spatial references to
be "GIS Friendly", meaning use lon,lat when transforming points.

Adding tests for this new function.