Issue #27 Exposing osr axis order and refactoring
Removing conditionals around setting axis order, as they are not
actually needed. GDAL 3 will properly set axis orders on the CRS if
geographic or projected.

Exposing the axis order with a constant at the top of utils that is used
and set to a default in def create_coordinate_transformation