Issue #6 adding back in table checks
Adding to validation that of checking threats are accounted for in
sensitivity table. Adding that same check back to the main execute
function as well as checking required sensitivty headers because these
are not to be assumed done by validation. For example if someone is
executing this from a python script.

Smaller name changes and condensing of variables