Issue #6 change resolving of threat rasters, tests
Moving the raise error out of the resolve threat raster path function
and handling it in the main execute function. This allows for better
error handling messages without needing to pass multiple things to the
function and keeps the function more generic.

Adding some tests and updating tests for this change