Check for threat values 0<=x<=1, add test
Refactoring handling of threat values where before values would be set
to 0 if less than 0 and to 1 if great than 1. Now checking 0 to 1 range
and throwing a ValueError if values outside that range. Adding a test
function for this behavior.

Big updates of tests in general to properly report and check for results
when threat values are floating point values. The regression values hard
coded in the tests come from running the tests and eye balling the
results. Also ran this version of the code against the sample data and
compared it to the results from InVEST 3.8. The results were identical.
Thus trusting these test data values are correct.

Still getting an error in one test for when threat raster has nodata
value of None. This might be a bug in ``convolve_2d`` and has been
reported.