removing unnecessary numpy where call
In computing habitat quality op there was a line that was setting
anything less than 0 from degradation sum to 0. But looking at the
user's guide and the equation, I don't think degradation should be less
than 0 by definition.

So it was unclear to me why this would need to be called.