Issue #122 refactoring operation for masking
There was a type bug in an operation that generates a 1, nodata mask of
infrastructure inputs. The created raster was set to GDT_BYTE but was
sent a nodata value of -1, causing that value to be set to 255.

Also, if only one infrasctucture raster was input there was a boolean to
integer type mismatch causing improper masking of nodata values.

Fix was to make the output mask raster integer type and be more explicit
about masking in the operation, using boolean masks on integer numpy
arrays.