I. #146 re-work of build lookup function
To maintain backwards compatible re-working function so that case
handling is done before any data is accessed.

Adding lots of tests for the changes made to the function including new
capabilities.

Fixing a bug in recreation where unique column was not being passed in
by the csv but was sneakily being overwritten.