Refactoring to better handle the results vector.
This (fairly large) commit updates how UFRM creates the output vector
and how we end up writing data to it depending on whether or not the
built infrastructure vector is present.  The starting point was pretty
complicated (and unnecessarily so, I think), and so I simplified it,
broke a few operations out into their own tasks and made the tasks a bit
more orthogonal.  RE:#168