Updating energy savings calculations.
This updates the UCM energy savings calculations to account for the
change in the consumption units to be not per building (of this class)
but to be kWh/deg C/m^2 for each building in the class.  This allows us
to account for classes to have defined heights, for example, or other
stranger qualities that could affect their energy consumption.

RE:#138