Making litter accumulation consistent with other pools.
Litter accumulation now has an initial value, and stocks accumulate
according to the rates of accumulation for litter for the current
lulc transition.  This is a departure from how litter was handled
in the previous version of the model, where the initial litter
value was used for each transition and that's about it.  RE:#209.