I. #8 add utils function and use in swy
Add utils function that takes a raster path band and dictionary and
returns values found in the raster that are not keys in the dictionary.
This functionality is useful for InVEST models that have a raster and
table input where the raster values must be present in a column of the
table.

Seasonal water yield now uses that functionality to test lulc values in
the biophysical table and to check soil group raster values are in the
set of 1-4.