I. #8 HQ test results update from convolve change
This is a long story. With the PGP 2.1 release a bug was introduced for
convolution. 2.1 was needed for this PR for
`ReclassificationMissingValuesError` addition to `reclassify_raster`.
This convolution bug was fixed in a post revision of 2.1 and to test it
I used Habitat Quality. However, HQ was using an invalid call to
`convolve_2d` ( see issue #270 ). Updating HQ caused HQ tests to be
different, hence this test result refactor.