I. #8 fixing ucm test due to pgp convolve changes
Using the latest 2.1+ version of `pgp.convolve_2d`, the nodata handling
is different. For the UCM test, the convolved raster was left with an
undefined nodata value which caused the `wbgt_op` to crash. Since this
is being worked on in PR #269 I copied the current solution from there
to hopefully avoid merge conflicts in the future.