I. #397 replacing dot w/ underscore for fieldnames
Updating the UG revision to reflect the changes to the UG that go with
this change. I realize this is a bit of a problem since UG/main has not
been merged with UG/3.9, so the UG will be out of wack. But that'll be
brought back in line when the 3.9 release is out.