Forcing environment variable defn, add warning.
The QT_MAC_WANTS_LAYER environment variable, though defined in
`exe/hooks/rthook.py`, does not appear to be used by the application at
runtime.  To counter this, I've added a definition for this environment
variable to the app bundle launcher shell script and also written a
runtime warning that is presented to the user on mac if the variable is
not already defined.  RE:#384