Removing experimental CDTP calls.
It turns out that the Chrome DevTools Protocol version 1.3 (the latest
officially released) does not contain the functionality needed to
control the window like I need it to for this test.  The experimental,
latest tip of development, does, but I don't want to have a test rely on
experimental functionality so I'm dropping it.

RE:#74