Removing logging from the execution thread for models.
Now that utils.prepare_workspace is more ubiquitous across InVEST, the
Qt UI is the one case where the execution.Executor thread does not log
the exception and function completion; these are already handled by
prepare_workspace.  There are several other use cases, however, such as
the datastack archiving functionality and validation, where we FOR SURE
do want to log exceptions, so this is now implemented as a conditional
in the Executor thread.  I've also made an effort to test this logging
via a QueueHandler in the tests, but the logging for the execution
completion when an exception _is_ raised does not seem to be recorded by
the QueueHandler and I don't know why.  Since we're deprecating the Qt
UI in the foreseeable future, this is probably not a huge deal.

RE:#651