Restoring our old osgeo hook.
The latest build logs for pyinstaller say that pyinstaller is failing to
locate `osgeo._gdal`, yet `_gdal.pyd` is available in the pyinstaller
CWD.  Our prior release of InVEST (3.9.0), which DOES have this hook,
includes the osgeo directory which contains a single file, `_gdal.pyd`.

So I'm hoping that this reinclusion of the hook will mean the
application will import `_gdal.pyd` as expected.

RE:#655