Refactoring greenspace budgets into raster_calculator.
After some thought and discussion with @dcdenu4, it seemed best to
refactor the iterblocks-based approach into 2 raster_calculator calls.
The main reasons to use iterblocks are if we either 1) need access to
the spatial context of the pixels or 2) need to modify a raster
in-place.  We need neither of these, and while it might save us some
runtime to avoid reading a raster from disk, this does seem the more
maintainable approach.

RE:#722