I. #835 fixing pandas df bug, update history
With pandas 1.4.0 a ValueError was being raised from the creation of a
pandas df. After reviewing the call it seems that the df has been
creating improperly. The df has been taking a map object iterable for
the columns argument and failing to create the desired columns. I
believe this was a uncaught exception in pandas before which is now
being properly handled in 1.4.0. However, because of how the rest of the
function uses that df, the columns are created and added anyway such
that no changes need to be made to the tests.