Implementing alignment.
Alignment in HRA is different because we can't guarantee that the user
will provide any rasters at all.  If only vectors are provided, it's all
too easy to rasterize and THEN align, which will lead to an unfortunate
degree of imprecision in the location of rasterized pixels.  The best
solution is to align the raster stack and rasterize the simplified
geometries onto the same grids, which is what this approach does.

RE:#819