Reworking the override system for datastack archives.
After some reflection (and when starting to prototype this), I realized
that it makes more sense to allow specific overrides of specific inputs.
In HRA, for example, only the criteria table needs to be overridden ...
all the other inputs are already described by ARGS_SPEC.  Why not just
provide the one special case that needs it?  So that's what I'm trying
to do now. RE:#819