Adjusting greenspace/population ratio.
TODO: check with the science team about correct behavior here.

This commit sets the greenspace/population ratio to the greenspace area
when the greenspace is less than 1.  The operative question here is what
to do when we have fractional people.  If fractional people are allowed,
the greenspace/population ratio will become very large, approaching
infinity as the population approaches 0 from the right.  It is unclear
what the correct behavior here should be.

RE #722