Reverting UNA to the state of 0ff3baf.
Per science team request, we are now assuming that administrative units
are nonoverlapping, and the rasters produced by the split population
groups are for information only.  It is not necessary to handle overlapping
polygons, as the actual calculation in the output admin units vector does
not actually use the rasters and so inherently handles overlapping
geometries.  RE:#722