I. #1104 Add a document to capture HQ decay change
Introducing a new location in the repository to track major design
decisions. ADR stands for Architecture Decision Records and is seemingly
a growing standard some software projects are using to record major
decisions. This is most commonly used after an RFC process. Since a lot
of our RFC process is via Slack, Google Docs, or other methods because
we interact with non-developers (our science colleagues), I felt like
capturing the decision endpoint with a recap summary was the most
efficient process.