Updating kernels to match the design doc.
In the draft user's guide chapter, several search kernels are defined
for weighting travel distances.  At least one of these (gaussian) is
similar to other kernels used elsewhere in InVEST.  However, the
definition of this gaussian kernel is different from the others in
InVEST.  Technically speaking, 'gaussian' refers to a class of function,
not a specific function.  Thus, I've reworked the kernels to define all
the kernels in accordance with what's defined in the user's guide.

RE:#722