Attempting a fix and noting change in history.
I believe the issue is that the percentiles function requires that the
directory is empty (so it can delete it later) or does not yet exist (so
that it can create it and then delete it later). The directory passed to
the percentiles function was not empty.  This is now fixed by passing in
a directory that does not yet exist, within the existing temp directory
used by the rest of the function.

RE:#1189